Kuvarti/reduce sdl version

+++ lib/sdl2/test/testautomation_guid.c
@@ -2,9 +2,8 @@
  * GUID test suite
  */
 
-#include <SDL3/SDL.h>
-#include <SDL3/SDL_test.h>
-#include "testautomation_suites.h"
+#include "SDL.h"
+#include "SDL_test.h"
 
 #ifdef HAVE_STDINT_H
 #include <stdint.h>
@@ -70,9 +69,9 @@ upper_lower_to_bytestring(Uint8 *out, Uint64 upper, Uint64 lower)
 /* Test case functions */
 
 /**
- * Check String-to-GUID conversion
+ * @brief Check String-to-GUID conversion
  *
- * \sa SDL_GUIDFromString
+ * @sa SDL_GUIDFromString
  */
 static int
 TestGuidFromString(void *arg)
@@ -95,9 +94,9 @@ TestGuidFromString(void *arg)
 }
 
 /**
- * Check GUID-to-String conversion
+ * @brief Check GUID-to-String conversion
  *
- * \sa SDL_GUIDToString
+ * @sa SDL_GUIDToString
  */
 static int
 TestGuidToString(void *arg)
@@ -117,7 +116,7 @@ TestGuidToString(void *arg)
 
         /* Serialise to limited-length buffers */
         for (size = 0; size <= 36; ++size) {
-            const Uint8 fill_char = (Uint8)(size + 0xa0);
+            const Uint8 fill_char = size + 0xa0;
             Uint32 expected_prefix;
             Uint32 actual_prefix;
             int written_size;
